+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v1.9.0 (unreleased)
 FEATURES:
 - **New Data Source**: azapi_resource_list
+- **New Data Source**: azapi_resource_id
 
 ## v1.8.0
 ENHANCEMENTS: